A little more progress on ham menu
All the contents of the page have been separated in a new div to allow
better handling when menu is out. This has caused a million more bugs to
appear. Not sure if I'm going to stick with this version, but I'll at
least try to get rid of the bugs.